tmdb-movie-project/requirements.txt
pandas>=1.5
requests>=2.28
python-dotenv>=1.0
matplotlib>=3.7
tqdm>=4.65
